fix(examples.dart): initialize the reflector
fixes #309
